move from utils to vector_operations and add tests
